Move `package.json` helpers into their own file
